Update README.md
Added setup.

(
## Setup
In the folder you will find 2 .txt files. (id.txt and ip.txt)

The id.txt is the id, every player should have thier own ID (unless you want bugs ^w^). (Numbers only)

The ip.txt is the IP of the server that you are trying to connect to. (LocalHost:6969 or "Ip of Pc you are connecting to":6969)

## Starting Server
To start the server, run the file with the name "server.bat".

The server should now start and run on your machine. (all computing devices with JAVA)

## Starting Game
To start/run the game you need the game in steam and have steam running. (not game from steam)
After starting steam you can run the file with the name "Game.exe", this will launch the game on your device. (Windows only)
)